move ShardingSchema & GlobalRegistry to runtime package
