+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/backend/netty/client/response/mysql/MySQLResponseHandler.java
@@ -23,8 +23,8 @@ import io.shardingsphere.core.metadata.datasource.DataSourceMetaData;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.shardingproxy.backend.netty.client.response.ResponseHandler;
 import io.shardingsphere.shardingproxy.backend.netty.future.FutureRegistry;
-import io.shardingsphere.shardingproxy.config.GlobalRegistry;
 import io.shardingsphere.shardingproxy.runtime.ChannelRegistry;
+import io.shardingsphere.shardingproxy.runtime.GlobalRegistry;
 import io.shardingsphere.shardingproxy.transport.mysql.constant.CapabilityFlag;
 import io.shardingsphere.shardingproxy.transport.mysql.constant.ServerInfo;
 import io.shardingsphere.shardingproxy.transport.mysql.packet.MySQLPacketPayload;
