+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/runtime/metadata/ProxyTableMetaDataConnectionManager.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.shardingproxy.config.metadata;
+package io.shardingsphere.shardingproxy.runtime.metadata;
 
 import io.shardingsphere.core.metadata.table.executor.TableMetaDataConnectionManager;
 import io.shardingsphere.shardingproxy.backend.jdbc.datasource.JDBCBackendDataSource;
