+++ sharding-proxy/src/test/java/io/shardingsphere/shardingproxy/runtime/metadata/ProxyTableMetaDataConnectionManagerTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.shardingproxy.config.metadata;
+package io.shardingsphere.shardingproxy.runtime.metadata;
 
 import io.shardingsphere.shardingproxy.backend.jdbc.datasource.JDBCBackendDataSource;
 import org.junit.Test;
